Deleted js-tree componenet as it was unecesary and generating error as well as confusions. Instead I created a class JsTree that contains helpful method to handle a jsTree from an html element. This class is used by tree tag componenet. Implemented refresh button which does not generate errors anymore.
