--- historian/frontend/src/app/shared/js-tree/js-tree.component.html
@@ -1,1 +1,0 @@
-<div #dataTree></div>
