--- historian/frontend/src/app/shared/js-tree/js-tree.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { JsTreeComponent } from './js-tree.component';
-
-describe('JsTreeComponent', () => {
-  let component: JsTreeComponent;
-  let fixture: ComponentFixture<JsTreeComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ JsTreeComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(JsTreeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
