+++ historian/frontend/src/app/tag/tag-tree/tag-tree.component.html
@@ -17,5 +17,4 @@ Select a TAG or multiple tags to edit or add to collection
     REFRESH TREE
 </button>
 
-
-<app-js-tree [jsonData]="treeDataTag"></app-js-tree>
+<div #dataTree></div>
